remove array_go18
