+++ array.go
@@ -3,6 +3,7 @@ package parquet
 import (
 	"unsafe"
 
+	"github.com/parquet-go/parquet-go/internal/unsafecast"
 	"github.com/parquet-go/parquet-go/sparse"
 )
 
@@ -20,4 +21,27 @@ func makeArrayString(values []string) sparse.Array {
 func makeArrayBE128(values []*[16]byte) sparse.Array {
 	ptr := *(*unsafe.Pointer)(unsafe.Pointer(&values))
 	return sparse.UnsafeArray(ptr, len(values), unsafe.Sizeof((*[16]byte)(nil)))
+}
+
+func makeArray(base unsafe.Pointer, length int, offset uintptr) sparse.Array {
+	return sparse.UnsafeArray(base, length, offset)
+}
+
+func makeArrayOf[T any](s []T) sparse.Array {
+	var model T
+	return makeArray(unsafecast.PointerOf(s), len(s), unsafe.Sizeof(model))
+}
+
+func makeSlice[T any](a sparse.Array) []T {
+	return slice[T](a.Index(0), a.Len())
+}
+
+func slice[T any](p unsafe.Pointer, n int) []T {
+	return unsafe.Slice((*T)(p), n)
+}
+
+type sliceHeader struct {
+	base unsafe.Pointer
+	len  int
+	cap  int
 }
